internal/pipe: get the package to build on Windows, too
